requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altgraph==0.17.5
annotated-doc==0.0.4
anyio==4.12.1
attrs==26.1.0
av==17.0.0
brotli==1.2.0
cachetools==5.5.2
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
ctranslate2==4.7.1
faster-whisper==1.2.1
filelock==3.25.2
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
g4f==0.5.7.0
google-api-core==2.30.0
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.73.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.7.2
idna==3.11
Jinja2==3.1.6
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.1
nest-asyncio==1.6.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.3
oauthlib==3.3.1
onnxruntime==1.24.4
openai-whisper==20250625
packaging==26.0
pefile==2024.8.26
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycryptodome==3.23.0
Pygments==2.19.2
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.3
pyparsing==3.3.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.18.0
pywin32-ctypes==0.2.3
PyYAML==6.0.3
QtAwesome==1.4.0
QtPy==2.4.3
regex==2026.2.28
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.3
rsa==4.9.1
shellingham==1.5.4
sympy==1.14.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
typer==0.24.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
yarl==1.23.0
